cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skillsctl"
version = "0.1.0"
description = "CLI tool for installing skills, agents, workflows, and rules from an Enterprise Skills Catalog"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Enterprise Skills Catalog"},
]
keywords = ["skills", "catalog", "cli", "enterprise", "agents", "workflows"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
skillsctl = "skillsctl.main:cli"

[project.urls]
Homepage = "https://github.com/your-org/enterprise-skills-catalog"
Repository = "https://github.com/your-org/enterprise-skills-catalog"
Issues = "https://github.com/your-org/enterprise-skills-catalog/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/skillsctl"]
